+++ inst/tests/test-family.R
@@ -1,8 +1,6 @@
 context("Family")
-roc <- rd_roclet()
-
 test_that("long families are wrapped", {
-  out <- roc_proc_text(roc, "
+  out <- roc_proc_text(rd_roclet(), "
     #' Title
     #' @family Long family name
     long_function_name_________________________1 <- function() {}
@@ -23,4 +21,20 @@ test_that("long families are wrapped", {
   seealso <- get_tag(out, "seealso")$values
   expect_equal(str_count(seealso, "\n"), 2)
 
+})
+
+test_that("special names escaped in family tag", {
+  out <- roc_proc_text(rd_roclet(), "
+    #' Title
+    #' @family Long family name
+    f <- function() {}
+
+    #' Title
+    #' @family Long family name
+    '%+%' <- function(a, b) {}
+  ")[[1]]
+  
+  seealso <- get_tag(out, "seealso")$values
+  expect_match(seealso, "\\\\%\\+\\\\%")
+  
 })